Avoid walking the same node twice
Because of the way declarations are handled, the same type node can be
attached to multiple variables, e.g.

int a, b;

will link the same 'int' type to both 'a' and 'b', so you get:

  decl +-> a --v
       |      int
       +-> b --^

So, you'll see the 'int' object twice -- once when from 'a' and
a second time from 'b'. Currently, it doesn't make any difference,
but the seen field may become important when 'a' and 'b' should
be of different types, because then the declaration would have
to be broken up into two.