Store pointer to the name/string dynstr in node_t
We had a char pointer for that purpose (or, rather multiple pointers
depending on the node type). But it turns out that having a shortcut
into the target dynstr is a bad thing, because it is no longer possible
to track the references. This is a problem if you try to replace the
typedef dynstr in constructs like:

off_t a, b;

The "off_t" dynstr is referenced by both "a" and "b", but there is
no way to know that from the tree representation.