Store the first basic type dynstr as node->str
This is not needed to get a correct parsed tree, but it helps greatly
simplify type replacement, because we can use the same infrastructure
as for other types.

This means that the refcounting always happens on the first dynstr
of a basic type, as if it were the actual type name.