Add split_list field to node_t
It is impossible to maintain the split list correctly if nodes are
taken off its original list and put on the split list, because the
position in the parent is lost.

The only option is a reparse of the whole decl (which is done in
type_split), but this is not easily possible while still walking
the tree.

So, adding a dedicated list_head seems like the least painful option.