Treat gdb print option variables as GDB types
Find variables that are assigned values from a call to
gdb_user_print_option_address(), and treat them as GDB
types.

This is a bit of a hack, because we should use the actual
type of the GDB internal variable, not what we decided them to
map to for GDB_COMMON, but the latter works just fine for "int"
and we don't have any such non-int variables.