Simplify and generalize find_scope
What we really want to have is a function that can return the scope
of any node, not only a containing function. Incidentally, all
such cases are covered by having a "body" element:

- function definitions have a chd_body list
- struct, union and enum declarations have a cht_body list

I'm not aware of any other possible contexts, but haven't read
the C99 specification either, so maybe there is. It should be added
here if it's needed...