Add rudimentary support for pointers
This includes pointers to functions and functions returning pointers.
However, no check is done that the level of indirection matches. The
assumption is made that there is no code like this:

long *p;
long x;
x = p;

Additionally, the compiler will issue a warning for this case.

However, this code will HAVE TO be improved if we add support for
typecasts.