subst_target_var: substitute all var duplicates, too
This is needed to solve situations like this:

static struct local_unwind_table {
        struct {
                unsigned long pc;
                unsigned long range;
        } core, init;
        void *address;
        unsigned long size;
} *local_unwind_tables, default_unwind_table;

In this case the "size" member is duplicated for each of the variables
(local_unwind_tables and default_unwind_table), but we don't want to
split this type into two structures, because they would become two
distinct structs. Instead, find the second copy of the variable and
change it too.