track_vars: also walk the splitlist
Some variables may not have been replaced yet, because they're still on
the split list. Their dependencies would be missed.

On the other hand, the tracking may add new entries to the replacedlist
(if no split is necessary), so we must walk both lists again until no
new entries are added to the replacedlist.