Add support for replacing pointer types
This code allows replacing pointer types for the first time.

We have to operate on the base type first (because that's the only
thing that can be omitted in a declaration of multiple vars), and
the original type has to be "flattened", e.g. all the pointer type
objects must be removed.

If you have e.g.:

void **p, *x;

and identify "void *p" as a candidate for replacement to "tptr", then
the shared part between p and x is "void" (the base type). But if you
decide to replace "void *" with "tptr", then you must remove the
extra asterisk to get:

tptr *p, x;