Never remove splits in replace_type()
There is a nasty interaction with type_subst(), when it walks over the
split list to track the just-to-be-substituted variables. If another
substitution is found and it is the last from the list, the split gets
removed in replace_type(). But the loop in type_subst() still holds
a stale pointer to a substituted type node (and split element). The
loop cannot continue.

To solve this, never ever remove elements from the split list in
replace_type. Instead, add extra code to type_split() to handle
the the case when all variables are replaced.