Re-initialize replacedlist after the final split
If this initialization is omitted, then the list gets corrupted, when
user data is reset for the next patch.

Now, you may question whether we can simply discard this information.
The fact is that the replacedlist list is (currently) only used to get
a list of variables that should be tracked. But since we also track
anything that's on the split list, there cannot be any replacements
that we haven't seen yet.