Fix varscope_find() API
The varscope_find API is flawed, because it's impossible to use it for
variable searches. If you already have an nt_var object, then you don't
need to call it, and if you only have an nt_expr object, then the API
wouldn't let you specify the nt_var type.