Track functions assigned to a function variable
Usually, we only adjust the left side of an assignment. With function
pointers, let's have a look at the right side as well, because in these
cases the right side is usually trivial.

You might actually want to add a complete top-to-bottom expression
tracker, but that would be a lot of effort with little reward...