Add support for a builtin file
The builtin file is a fake file that contains declarations that should
be added to the internal declarations. This has been faked by
predef_types for the parser, but sometimes we need more information,
e.g. to track a struct element.

The builtin file can be recognized by a NULL name.