Fix flatten_type for abstract type names
In a declaration, the type's dynstr is part of the declarator, and the
abstract type does not contain it. However, for a type_name element
(e.g. in sizeof() or typecasts), the resulting type covers the complete
expression. For example, in:

foo = (char *) bar;

the "char *" pointer type contains 3 dynstrs: "char", " " and "*". When
you try to flatten this type, you'll try to delete the pointer node,
but this fails, because there is still a reference to "char" from the
base type. This is in fact correct, because otherwise the resulting
flattened type would contain a dangling pointer to an already freed
dynstr.

Fix this by adjusting the abstract type's extents before deleting it.