track_expr: save indirection specifiers when dereferencing
The indirection stack is passed by reference, so if you pop some
values from it, they are not restored, but the calling function may
still hold a pointer to them.

To fix this with a minimum of code (and without breaking tail recursion
optimizations in most places), let's save the values which are popped,
and return them after a recursive call to track_expr().