target_facilitators: use variable tracking for the initial substitution
The variable tracking code is more robust than the find_assign() hack.
For example this code would confuse the old code, but works fine with
the new approach:

char *buf = GETBUF(ULONG(hdr));

The left side of this assignment would be mistakenly treated as a target
type, because the call to GETBUF() was not properly analyzed.