Change the meaning of "global" scope
Since different files in the project contain different definitions for
the same type name, it is not possible to treat file scope as global,
even for types.

In any case, looking at the linkage attributes is wrong, because this
is only relevant to the linker, and we don't really care how objects
are linked.

Let's do it differently: if something is declared in a header file, it
is assumed to be included in other files as well. This is the GLOBAL
scope now. Anything else is assumed to be used only from the file itself
and gets a file scope.

Yes, I know it's still a hack - ideally we would track which file
includes what and duplicate each included file as necessary, effectively
removing the need of a "global" scope altogether.