Increase the hash table size
Currently, there are over 25000 objects in the varscope hash. Even though
the hash function has improved, the most typical chain length is 6
(corresponding to the total number/hash size ratio).

With 8192 the most common chain length drops down to 2, which is quite
good. FWIW increasing the table size to 16384 doesn't seem to make the
situation much better for several reasons, but this observation might
be reviewed later if we measure the actual lookup performance.