varscope: Find global declarations for file-level scope
This still allows you to have different file-level declarations in
different files, but if there is a global declaration as well, it must
be compatible. The idea is that header files are included from every
source file, so you can't have a file-level declaration that conflicts
with a global one.