Reverse ind_t stack direction
Let it grow downwards. This makes debugging a bit easier, because gdb
interprets the stack as a C string. Since the stack always ends with
ind_stop, which is numerically identical to the NUL character, you can
directly see the whole stack in gdb if it is ordered this way.