is_host_type: treat macro expansions as non-determinate type
Using a macro without parameters is parsed the same as a function
identifier, which would normally be a host-specific function pointer
type. A macro expands instead.

TODO: It would be better to trace the macro expansion, but that would
be a major effort. Better fix the result afterwards if necessary.