track_args: do not track pointers to unconverted types
This is intended to help with cases like free(ptr), where 'ptr' can
be either a pointer to target-specific data or a pointer to
host-specific data.

I could have made a blacklist, but it turns out that this approach
is just as effective (for our code base) and requires less maintenance.