Add functions for rolling back node from replacedlist
Sometimes you don't know whether a change should be done until you try
to make the change. But if the substitution shouldn't have been done,
there must be a way to roll back the effects.

For the replacedlist, it boils down to removing any new nodes from the
end of the list. Despite the "cool" name of the functions. ;-)