parser: set type_pointer node extents to the '*'
This is consistent with the handling for variables.

Most importantly, flatten_type() fails miserably if we set the extents
to the full type name, because it will try to remove the pointer node,
which also contains the base type dynstr, and that is still referenced
by the base type.

So don't do that.