get_cpp_cond: do not detach raw text
The raw text is now part of the parsing, so there's no longer any
need for the detach quirk. That one was only needed to avoid discarding
the raw text later in the parser (and thus invalidating any pointers
to the dynstr objects).

The old way was:

"#ifdef XYZ"  (one token, also written verbatim to the output)
"#ifdef" " " "XYZ"  (many tokens from the second parsing, not used)

The new way is:

"#" "ifdef" " " "XYZ"  (used for the output and stored in get_cpp_cond)