Do not expand macro arguments during argument scanning
Argument substitution is just like macro substitution, but it must not
happen while we are still scanning the argument list, otherwise we would
try to expand the argument incorrectly.

Consider this:

#define FIRSTCHAR(s)  (s[0])
#define s             xyz

Then "FIRSTCHAR(s)" should be first expanded to "FIRSTCHAR(xyz)", even
though "s" happens to be the parameter name as well.