Free up macro arguments after expansion
Macro arguments are not needed after they are expanded inside the
corresponding macro, so they can be removed.

In fact, this is not merely a cleanup patch, because the argument
expansion may interfere with the expansion of a calling macro,
effectively re-using old text for a subsequent macro expansion.
If the macro argument text is discarded, this cannot happen, obviously.