Keep a copy of the macro parameter value before expansion
For stringification ('#') and concatenation ('##'), the unexpanded
value of the parameter is needed, so we must save both.

That means that there are always two hashed_macro objects for each
parameter, and because they have the same name, they also share the
same slot, hence the newarg (expanded) must point to the unexpanded
argument in its ->next pointer.