Implement CPP concatenation
Finally implement the last big missing piece from the pre-processor.

The idea is that dupconcat() creates a new dynstr object, which only
contains the two tokens pasted together (without any respective
whitespace) and this dynstr is then re-parsed using yylex().

As the last thing, the original tokens are replaced by the result
of the re-parsing. Note that the result is also put into the right
place, because dupds was initially inserted before the insertion
point, so we don't have to do anything about it when we call
replace_text_list().