Add BLOCK expression type
There is some need to mark compound statements, e.g. to set correct
variable scope (for declarations inside a block). Previously, the
following two codes would produce the same parse tree:

    statementA;
    statementB;
    statementC;

and:

    statementA;
    {
        statementB;
    }
    statementC;

After this change, statementB is put inside a BLOCK expression, which
is in turn added to the list with statementA and statementB.