is_host_type: ignore the type of a typecast
The target typecast type is not relevant to determine whether a variable
is host-specific or target-specific. In particular, this construct is
broken:

    uaddr_t *sbp, *asp;
/* ... */
    asp = (uaddr_t*)((uaddr_t)sbp + (STACK_SIZE - (saddr - sp)));

This first casts sbp to an integral type, so that the offset is
added as an absolute number and not multiplied by sizeof(uaddr_t).

Of course, it would be much cleaner to typecast sbp to a "char *"
instead, but...