Preserve macros when reparsing a node
This has two parts:

A. Mark the result of expansion as fake - this is done by storing the
   appropriate flags
B. Copy the unexpanded text verbatim - this is done by making a duplicate
   and adding it to the appropriate position in the output.

Implementation:
1. store the input meta-information in dynstr_read()
2. adjust the output in store_yytext()

Note that the lexer may combine the text before and after the macro
(e.g. if both is whitespace), and then it's necessary to split it again
in store_yytext().

Also note that the stored queue is always empty when reading from
a non-dynstr stream. In that case there is no change from the previous
behaviour.