cleanup: rename "expanded" to "macro"
The name of this flag is quite confusing, because it marks the macro
text before expansion. The idea was that the text was expanded into
something else, but that doesn't really convey the meaning very well.

While it could theoretically be used for other purposes, for now it
always marks the verbatim macro text, so let's call it simply "macro".