Fix stringification for re-parsing
After stringification, the resulting token is correctly set to
STRING_CONST, but the text is left as-is. This breaks re-parsing,
because all tokens are re-interpreted then.

Add the missing quotes to fix it.

NOTE: We should also escape all quotes inside the string, but this is
never needed for xcrash, so I'm leaving it out.