Trace back macro locations in yyerror()
To simplify understanding parser errors within macro expansion,
add a mechanism to trace back where and how the macro was used.

Simply add a "parent" location, which is set to the input for the
expansion. On an error, print all the parent locations, which should
give a pretty good idea of what went wrong.