mkstring_variadic: fix reparsing
The first argument is modified incorrectly. Although the resulting text
is correct, the underlying structures are all wrong:

  * No changes are made to the fake expanded text.
  * The argument's first_text starts after the actual text that is
    written to output, hence the typecast ends up in the wrong place
    (at the end of the argument, rather than at its beginning).

In fact, the dynstr chain claims that the first argument of MKSTR()
expands to the same text as before.

To fix this, remove the parsed node completely, remove all fake text
and remove the "macro" flag from the input. This can be done, because
the complete node is reparsed, so the parse tree needn't be preserved.