cleanup: get rid of fileloc
This hack should never have been created, but the clean solution is
a bit tricky, too.

Basically, we have to redefine Bison's "$end" symbol, because the
original name cannot be used in grammar (isn't there a reason for it?).
Next, the rule must include YYACCEPT, because we'll never get two
$end symbols from the lexer.

There might be a catch with this solution, but it seems to work. ;-)