parser: make the location argument non-optional
There must now always be a location associated with a parser. Rename
it to "fileloc", because it holds the extents of the complete file
on return.

However, note that it is still an IN/OUT parameter, so you must
initialize it first. This can be useful when reparsing a single node,
because you can get the actual line numbers in the parser now.