Kill the reuse flag
The reuse flag is broken by design, because we always want to reuse
the last object, unless we can't. We can't if:

A. the new string should be referenced by a node
B. the flags are different

None of these conditions can be determined when the previous dynstr
is read, because it depends on the type of the next dynstr.

So, let's store the yytext at the END of each flex action instead of
at the beginning, and make sure that if we read a token, it doesn't
get combined with either the previous or next dynstr.