parser: get rid of the empty dynstr at begin
This was needed to make sure that we have a valid first.text for
the file location. However, we can reuse the trick that is already
used for reparsing:

1. at begin, store the last dynstr in raw_contents
2. at end, set the first dynstr to the next token from the previous last

Note: if there are no dynstr objects in the chain, we'll get an invalid
struct dynstr pointer, whose list member overlaps raw_contents. It is
possible to use this condition to check for an empty list.