reparse_node: set up file location correctly
Since the parser extends the file location at end, and we're about
to reparse the complete expression, we should initially set end to
point to the beginning (i.e. empty).

If we don't do that, line numbering will be wrong for the reparse if
there is any newline in the input string.