lexer: mark macro name as ID token to prevent reusal
When a macro is used, the macro name is marked as an empty token, hence
subject to possible merge with the following whitespace. But if this
happens, the pointer in the corresponding location variable is not
updated. Since we cannot (easily) track the location pointer, it's
necessary to mark it as non-combining (non-empty token).