remove_text_list{,_rev}: optimize list operations
Apart from potential performance win (moving nodes from first_list and
last_list one by one somehow feels suboptimal), there is an improvement
in the usability.

Previously, the @keep parameter had to be a real, existing dynstr (not
e.g. the pseudo-dynstr at &raw_contents). With the change, the virtual
end-of-list marker can also be used as @keep, which allows deleting
from the beginning and/or end of the list.