remove_text_list: replace the removed text with a null dynstr
Do not use dummydynstr as a replacement, because that breaks many
assumptions about the continuity of the text list. Instead, create
a new empty dynstr which can be used as a replacement.  If the
replacement isn't used in the end, it is removed again.

Note that the do_remove() function is more general. This should allow
rewriting replace_text_list() in a more efficient way.