lexer: kill cpp_recover state
This state is almost identical to cpp_ign, but it doesn't terminate
the lexer. Since this is error recovery (i.e. not a hot path), we
can simplify the lexer itself even at the cost of a bit more runtime.

The trick is to call yylex() recursively, ignore the 0 return value
and continue scanning.