lexer: enforce caller's flags when reading macro args
There was an issue with preserving the flags when reparsing a node,
because the caller wants to mark all new content as macro, but we are
parsing content that may not be a macro.

We cannot keep the macro flag for the text to be reparsed, because
unexpanded macros are skipped when parsing from a dynstr list on the
assumption that each macro is followed by its expansion. Instead, we
ignore the stored flags when reading macros, because the macro arg
parser knows better what flags should be given to the result.

Note that this is only possible, because the cpp_arg state is "sticky",
that is, there are no transitions possible from the cpp_arg state.