Undo macro expansion in convert_readmem
When replacing the arguments to readmem(), we'd better preserve the
macros in the output, but the nodes only point to the expanded text.
However, it is possible to undo the effect of macro expansion, because
the expansion details are saved in the corresponding struct macro_exp.

Provide a utility function for doing that: unexpand_macros()