reparse_node: clean up CPP conditions
Clean up CPP condition syntax trees before cleaning up raw contents.
This is needed, because otherwise the corresponding CPP identifiers
are still referenced by the condition tree nodes, and freedynstr()
fails with "Attempt to free an in-use dynstr!".

Interestingly, this was not an issue before we switched to parsing
CPP directly from the lexer...