get_cpp_cond: avoid unnecessary node duplication
For #elif, the pre-condition is duplicated twice - once when getting
the last condition or when combining it with the pre-condition, and
once when adding the new condition.

The flaw is that state->precond may itself be used elsewhere, so it
cannot be added to the combined tree. But for #elif, a new node must
be created to combine it with the negation of the last condition.
So, we must treat #elif specially, without duplicating state->precond,
because in this special case, the pre-condition is already a duplicate.

Now, this discussion may seem too puristic, but the extra duplicated
nodes are not found in the raw list. Consequently, they are never freed,
but since they still hold a reference to the corresponding dynstr
objects, it is not possible to free the raw file content when a new
parsing is needed (e.g. after importing a patch).