cleanup: kill node->pf and parsed_file
The pf field in struct node duplicates the file from the location.

But if we remove that field, the "parsed_file" global is no longer
needed, because it was used solely to initialize that field.

Note that the first and last field of location must be equal, because
we never parse across file boundaries.