cleanup: duplist -> insert_dup_macro()
The duplist() function has been re-implemented in tools.c except the
NULL-pointer handling (empty macros). While rewriting the call sites,
it turned out that a bit more code can be reused, because we always
duplicate a macro at a given insertion point, hence the new function.

Note that this commit fixes a potential crash if a concatenated macro
expands to an empty string (NULL).