Fix dynstr flags after replacing strings in printf_spec_one()
The basic idea works, but different copies of the original string
should have different parts removed, depending on the original flags.

So, let's reparse the first node with null flags to get the template,
which may itself include macro expansion. Then adjust the resulting
string for each duplicate:

* For the original macro text, remove the expansion and mark all
  other items with the macro flag.
* For the expanded text, remove the original macro and mark all
  other items with the fake flag.

That produces the same results as parsing the changed text.